feat: using tdd, ensure category and ingredient exists before deleting

+++ Grubly/Services/CategoryService.cs
@@ -34,6 +34,13 @@ public class CategoryService : ICategoryService
 
     public async Task DeleteCategory(int id)
     {
+        Category? existingCategory = await _categoryRepository.GetOne(id);
+
+        if (existingCategory == null)
+        {
+            throw new KeyNotFoundException($"Category with ID: {id} not found.");
+        }
+
         await _categoryRepository.Delete(id);
     }
